frontend/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.5
xai-sdk==1.5.0
xdk==0.4.6
